requirements.txt
websocket-client
python-osc
python-dotenv